# apps/api/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: PBX Control Plane API
  version: 0.1.0
  description: >
    Multi-tenant PBX/UC platform control plane. All tenant-scoped endpoints are
    isolated by the authenticated user's tenant_id; the tenant id is taken from
    the validated JWT, never from request body or query.
    A live, generated spec is served at /docs by @nestjs/swagger. This file is a
    hand-written reference subset (auth/login, tenants, extensions, cdr) kept in
    version control for review and onboarding.
servers:
  - url: http://localhost:4000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: extension not found
            details:
              type: object
              additionalProperties: true
...
    Tenant:
      type: object
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: Acme Telecom }
        domain: { type: string, nullable: true, example: pbx.acme.com }
        plan: { type: string, example: pro }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /tenants:
    get:
      tags: [tenants]
      summary: List all tenants (platform admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: All tenants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tenant' }
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /tenants/me:
    get:
      tags: [tenants]
      summary: Get the caller's own tenant
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The authenticated user's tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
...